Add files via upload
The weather dataset is a time-series data set with information recorded hourly about the weather conditions at a particular location. Some of the metrics recorded are Temperature, Apparent Temperature, Precip Type, Humidity, Wind Speed, Wind Bearing, Visibility, Pressure and Daily Summary.
This data is available as a .csv file, we are going to analyse this data set using Pandas Dataframe.